cleanup captplanet config
