+++ cmd/captplanet/setup.go
@@ -17,24 +17,6 @@ import (
 	"storj.io/storj/pkg/provider"
 )
 
-// SetupCaptplanet defines Captain Planet setup configuration
-type SetupCaptplanet struct {
-	SatelliteCA         provider.CASetupConfig
-	SatelliteIdentity   provider.IdentitySetupConfig
-	UplinkCA            provider.CASetupConfig
-	UplinkIdentity      provider.IdentitySetupConfig
-	StorageNodeCA       provider.CASetupConfig
-	StorageNodeIdentity provider.IdentitySetupConfig
-	ListenHost          string `help:"the host for providers to listen on" default:"127.0.0.1"`
-	StartingPort        int    `help:"all providers will listen on ports consecutively starting with this one" default:"7777"`
-	APIKey              string `default:"abc123" help:"the api key to use for the satellite"`
-	EncKey              string `default:"insecure-default-encryption-key" help:"your root encryption key"`
-	Overwrite           bool   `help:"whether to overwrite pre-existing configuration files" default:"false"`
-	GenerateMinioCerts  bool   `default:"false" help:"generate sample TLS certs for Minio GW"`
-
-	Captplanet
-}
-
 var (
 	setupCmd = &cobra.Command{
 		Use:         "setup",
@@ -42,12 +24,12 @@ var (
 		RunE:        cmdSetup,
 		Annotations: map[string]string{"type": "setup"},
 	}
-	setupCfg SetupCaptplanet
+	setupCfg Captplanet
 )
 
 func init() {
 	rootCmd.AddCommand(setupCmd)
-	cfgstruct.Bind(setupCmd.Flags(), &setupCfg, cfgstruct.ConfDir(defaultConfDir))
+	cfgstruct.BindSetup(setupCmd.Flags(), &setupCfg, cfgstruct.ConfDir(defaultConfDir))
 }
 
 func cmdSetup(cmd *cobra.Command, args []string) (err error) {
